pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ml_classifier"
version = "0.1"
requires-python = ">=3.12"
dependencies = [
    "flask==2.2.5",
    "gunicorn==23.0.0",
    "jax==0.10.0",
    "jaxlib==0.10.0",
    "keras==3.14.0",
    "numpy==2.1.3",
    "pandas==2.2.3",
    "PyYAML==6.0.2",
    "scikit-learn==1.6.1",
]

[tool.setuptools.packages.find]
where = ["."]
